+++ app/src/main/java/com/example/solar_alarm/Data/Repositories/SolarTimeRepository.java
@@ -4,6 +4,7 @@ import android.app.Application;
 
 import androidx.lifecycle.LiveData;
 
+import com.example.solar_alarm.Data.Daos.SolarTimeDao;
 import com.example.solar_alarm.Data.SolarAlarmDatabase;
 import com.example.solar_alarm.Data.Tables.SolarTime;
