+++ app/src/test/java/com/example/solar_alarm/LocationTests.java
@@ -5,7 +5,7 @@ import android.content.Context;
 import androidx.room.Room;
 import androidx.test.core.app.ApplicationProvider;
 
-import com.example.solar_alarm.Data.Repositories.LocationDao;
+import com.example.solar_alarm.Data.Daos.LocationDao;
 import com.example.solar_alarm.Data.SolarAlarmDatabase;
 import com.example.solar_alarm.Data.Tables.Location;
